% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpiechartclustermarkers.R
\name{addpiechartclustermarkers}
\alias{addpiechartclustermarkers}
\title{Change leaflet cluster markers to pie charts}
\usage{
addpiechartclustermarkers(map, .data, .colors, group)
}
\arguments{
\item{map}{the map to add awesome pie chart cluster markers to}

\item{.data}{data for the cluster markers}

\item{.colors}{a vector of colors of at least the same size that nlevels(.data[[group]])}

\item{group}{the name of a factor variable of .data}
}
\description{
Change leaflet cluster markers to pie charts
}
\examples{
data("breweries91",package="leaflet")
breweries91$goodbear<-sample(as.factor(c("terrific","marvelous","culparterretaping")),nrow(breweries91),replace=T)
library(leaflet)
library(dplyr)
leaflet() \%>\%
addTiles() \%>\%
addpiechartclustermarkers(breweries91,.colors=c("red","green","blue"),group="goodbear")
}
