% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{backtest}
\alias{backtest}
\title{Backtest Strategy}
\usage{
\S4method{backtest}{Strategy}(object, horizon = "6m", data.width = "24m",
  keep.history = F, optim.param = NULL, optim.param.min = 1,
  optim.param.max = 10, optim.param.scale = 0.1, from = NULL,
  until = NULL, which = NULL, rf = 0, printSteps = F)
}
\description{
Backtest Strategy
}

