% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{closestpointsontwosegments}
\alias{closestpointsontwosegments}
\title{computes the coordinates of the closest point on a segment to a point in the plane}
\usage{
closestpointsontwosegments(s1, s2)
}
\arguments{
\item{s1}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}

\item{s2}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}
}
\description{
computes the coordinates of the closest point on a segment to a point in the plane
}
