% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{dist_areas_f}
\alias{dist_areas_f}
\title{Distances between hexagonal bins}
\usage{
dist_areas_f(
  U,
  delta = (range(U$x)[2] - range(U$x)[1])/100,
  h = neighbourhoods(U, delta)
)
}
\arguments{
\item{U}{: a dataframe containing the numerical variables x and y and preferable hexagon}

\item{delta}{: needed if hexagon is not a variable of U: bins will be recomputed}
}
\value{
a named matrix
}
\description{
Distances between hexagonal bins
}
\examples{
data(U) 
dist_areas_f(U)

delta<-0.01
h<-neighbourhoods(U,delta)
U$hexagon<-paste0(h@cID)
hD<-dist_areas_f(U,h)

sss1=sample(nrow(U),1000)
sss2=sample(nrow(U),1000)
x=sapply(1:1000,function(i){dist(U[c(sss1[i],sss2[i]),c("x","y")])})
y<-sapply(1:1000,function(i){hD[U$hexagon[sss1[i]],U$hexagon[sss2[i]]]})
plot(x,y,pch=".")
}
