% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{distpointtoseg}
\alias{distpointtoseg}
\title{computes the distance between a point and a segment}
\usage{
distpointtoseg(p, s)
}
\arguments{
\item{p}{a numeric vector of length 2}

\item{s}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}
}
\description{
computes the distance between a point and a segment
}
\examples{
zz<-function(p){
s=matrix(c(0,3,0,0),2,2)
plot(s,type="l",xlim=c(-2,5),ylim=c(-2,2))
points(x=p[1],y=p[2] ,col="red",cex=.5)
points(projpointonseg(p,s)[1],projpointonseg(p,s)[2],col="red",cex=.5)
segments(x0 = p[1],y0=p[2],x1=projpointonseg(p,s)[1],y1=projpointonseg(p,s)[2],col="red")
text((projpointonseg(p,s)[1]+p[1])/2,(projpointonseg(p,s)[2]+p[2])/2,round(distpointtoseg(p,s),2))}
par(mfrow=c(3,3))
set.seed(1);replicate(9,zz(c(sample(-2:3,1),sample(-2:2,1))))
}
