% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{distpolytopoly2}
\alias{distpolytopoly2}
\title{Computes distance between two polygons, only works for polygons that do not intersect}
\usage{
distpolytopoly2(poly1, poly2)
}
\arguments{
\item{poly1}{a polygon (a n x 2 numerical matrix)}

\item{poly2}{a polygon (a n x 2 numerical matrix)}
}
\value{
a positive number, the distance between the two polygons
}
\description{
Computes distance between two polygons, only works for polygons that do not intersect
}
