% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{distsegmenttopoly}
\alias{distsegmenttopoly}
\title{Distance between a segment and a polygon}
\usage{
distsegmenttopoly(s, .poly)
}
\arguments{
\item{s}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}

\item{.poly}{a polygon (a nx2 matrix, each line is a point)}
}
\description{
Distance between a segment and a polygon
}
\examples{
data(Avo_fields,package="Strategy")
polygon1<-Avo_fields[1,]
A<-polygon1@polygons
B<-A[[1]]@Polygons[[1]]@coords
s<-cbind(runif(2,min = min(B[,1]),max=max(B[,1])),runif(2,min=min(B[,2]),max=max(B[,2])))
plot(B,type='l')
s1<-s
points(s,type="l",lwd=4,col="green")
x<-vector()
for(i in 1:(nrow(B)-1)){
s2<-B[(i:(i+1)),]
dd<-distsegmenttosegment(s1,s2)
l<-which(c(distpointtoseg(s1[1,],s2),distpointtoseg(s1[2,],s2),distpointtoseg(s2[1,],s1),distpointtoseg(s2[2,],s1))==dd)[1]
min(as.matrix(dist(rbind(s1,s2),diag=T,upper = T))[3:4,1:2]);dd
sk<-if(l<=2){s2}else{s1}
p=rbind(s1,s2)[l,]
points(x=p[1],y=p[2] ,col="red",cex=2)
points(projpointonseg(p,sk)[1],projpointonseg(p,sk)[2],col="red",cex=2)
segments(x0 = p[1],y0=p[2],x1=projpointonseg(p,sk)[1],y1=projpointonseg(p,sk)[2])
x=c(x,dd)
}
min(x)
distsegmenttopoly(s,B)
distpolytopoly()
}
