% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{extractpolygonsaslist}
\alias{extractpolygonsaslist}
\title{converts a shapefile to list of  polygons (nx2 matrices)}
\usage{
extractpolygonsaslist(shp)
}
\description{
converts a shapefile to list of  polygons (nx2 matrices)
}
