% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{Generate_Constrained_Epidemic}
\alias{Generate_Constrained_Epidemic}
\title{Generate epidemic under size constraint}
\usage{
Generate_Constrained_Epidemic(
  U,
  .distriskhalf = 5 * 10^(-4),
  jumprisk = 10^-6,
  delta = 0.05,
  numberinfected = 10,
  foyersaleatoires = 2
)
}
\arguments{
\item{U}{a data.frame with x and y}

\item{.distriskhalf}{a positive number(default 5*10^(-4))}

\item{jumprisk}{=10^-6 a positive number}

\item{delta}{=0.05 a positive number}

\item{numberinfected}{target number of  infected elements of the population,}

\item{foyersaleatoires}{number of sources at random at the start}
}
\description{
Generate epidemic under size constraint
}
\examples{
.distriskhalf=5*10^(-4);jumprisk=10^-6;delta=0.05; TT=10
UE<-Generate_Discrete_Time_Epidemic(U,3)
}
