% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_U2.R
\name{generate_U2}
\alias{generate_U2}
\title{Generate a population of avocado trees}
\usage{
generate_U2()
}
\description{
Generate a population of avocado trees
}
\examples{
data(U2,package="Strategy")
data(Avo_fields,package="Strategy")
data(U2,package="Strategy")
U2<-U2[2:4]
#Plot the trees
Avo_fields$Source_yr<-addNA(as.factor(Avo_fields$Source_yr))
QLD<-Avo_fields$State=="Qld"
Avo_ids<-unique(Avo_fields[QLD,]$Avo_id)[c(1,3,4)]
QLD<-QLD&is.element(Avo_fields$Avo_id,Avo_ids)
QLDt<-is.element(U2$id,Avo_ids)
yearpal <- colorFactor(heat.colors(5),domain = levels(Avo_fields$Source_yr),na.color = "#aaff56")
leaflet(Avo_fields[QLD,]) \%>\%
addProviderTiles('Esri.WorldImagery',options = providerTileOptions(minZoom = 1, maxZoom = 21,maxNativeZoom=19)) \%>\% 
addProviderTiles("CartoDB.PositronOnlyLabels")\%>\% 
addPolylines(fillOpacity = 1, weight = 3, smoothFactor = 0.5,opacity = 1.0,
            color=~yearpal(Avo_fields[QLD,]$Source_yr),
                        fillColor=~yearpal(Avo_fields[QLD,]$Source_yr))\%>\% 
                        addMarkers(lng = U2[QLDt,]$x1,lat = U2[QLDt,]$x2,clusterOptions = markerClusterOptions())
}
