% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{getBacktestSetup}
\alias{getBacktestSetup}
\title{Get backtest parameter values from \code{Strategy}-object}
\usage{
getBacktestSetup(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Gets the backtest parameter values of an object of class \code{Strategy} that were used for backtesting the strategy. This includes the information about the parameters,
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get backtest setup from MA(200)-Strategy
getBacktestSetup(myStrat.MA)
}

