% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getCosts}
\alias{getCosts}
\alias{getCosts,Strategy-method}
\title{Get strategy function from \code{Strategy}-object}
\usage{
getCosts(object)

\S4method{getCosts}{Strategy}(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Returns the fixed and relative trading costs of an object of class \code{Strategy}..
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Get strategy function from MA(200)-Strategy
MA.costs <- getCosts(myStrat.MA)
# return fix costs
MA.costs$fix
# return relative costs
MA.costs$relative

##End(Not run)
}
