% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getFilters}
\alias{getFilters}
\alias{getFilters,Strategy-method}
\title{Get strategy values from \code{Strategy}-object}
\usage{
getFilters(object, which = NULL)

\S4method{getFilters}{Strategy}(object, which = NULL)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{which}{Which filters shall be returned. Either list number or names to be passed.}
}
\description{
Gets the strategy values of an object of class \code{Strategy} that were output from strategy calculation.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Get strategy values from MA(200)-Strategy
getFilters(myStrat.MA) # all strategy values returned

##End(Not run)
}
