% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{getParameters}
\alias{getParameters}
\title{Get strategy function from \code{Strategy}-object}
\usage{
getStratFUN(object, use.backtest=F)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{use.backtest}{If set to \code{TRUE}, the calibrated parameters of the backtest are returned. Requires \code{\link{backtest}} to be  executed first.}
}
\description{
Gets the strategy function of an object of class \code{Strategy} that was used for strategy calculation.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get parameters from MA(200)-Strategy
getParameters(myStrat.MA)
}

