% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{getStratFUN.name}
\alias{getStratFUN.name}
\title{Get strategy function name from \code{Strategy}-object}
\usage{
getStratFUN.name(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{include.params}{If set to \code{TRUE}, the parameters used for strategy evaluation are included.}
}
\description{
Gets the strategy function name of an object of class \code{Strategy} that was used for strategy calculation. This function is for aesthetical purposes only and does not have any numerical relevance.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get strategy function name from MA(200)-Strategy
getStratFUN.name(myStrat.MA) # returns "MA"
getStratFUN.name(myStrat.MA, include.params=T) # returns "MA(200)"
}

