% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{getStratFUN}
\alias{getStratFUN}
\title{Get strategy function from \code{Strategy}-object}
\usage{
getStratFUN(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Gets the strategy function of an object of class \code{Strategy} that was used for strategy calculation.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get strategy function from MA(200)-Strategy
MA.FUN <- getStratFUN(myStrat.MA)
}

