% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{getStratVals}
\alias{getStratVals}
\title{Get strategy values from \code{Strategy}-object}
\usage{
getStratVals(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Gets the strategy values of an object of class \code{Strategy} that were output from strategy calculation.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get strategy values from MA(200)-Strategy
getStratVals(myStrat.MA) # all strategy values returned
}

