% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonizepop.R
\name{hexagonize}
\alias{hexagonize}
\title{Create hexagons to cover an area}
\usage{
hexagonize(.spatialdata, hexagondiameterinangle = 5)
}
\arguments{
\item{U}{a spatial object  of class sf.}

\item{delta}{= a positive number:length in meters of the side of smallest hexagon.}
}
\description{
Create hexagons to cover an area
}
\examples{
library(sf)
.spatialdata<-st_as_sf(LocustAnalysis::Swarms)
hexagonize(.spatialdata) 
}
