% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{MDD}
\alias{MDD}
\title{Strategy Performance Maximum Drawdown}
\usage{
MDD(object, which=NULL, of="portfolio", from=NULL, until=NULL, type="relative"
     , use.backtest=F
     , ...)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{of}{Maximum Drawdown to be calculated for assets separately or the portfolio.}

\item{from}{The date in character format \code{"yyyy-MM-dd"} or as date-object from which performance shall be considered. If \code{NULL}, no restriction is made.}

\item{until}{The date in character format \code{"yyyy-MM-dd"} or as date-object until which performance shall be considered. If \code{NULL}, no restriction is made.}

\item{which}{Names or number of assets that should be included in calculation.}

\item{type}{If the \code{absolute} or \code{relative} drawdown of the performance shall be returned.}

\item{use.backtest}{If set to \code{TRUE}, the signals from the backtesting output are considered for maximum drawdown calculation. If \code{FALSE}, the signals from the initial strategy execution are used.}
}
\description{
Gets the maximum drawdown of the performance of an object of class \code{Strategy}.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get MDD of MA(200)-Strategy portfolio
MDD(myStrat.MA, from="2015-01-01", until="2015-12-31")

## Not run:
# Get MDD of MA(200)-Strategy (daily data = 252 trading days)
MDD(myStrat.MA, from="2015-01-01", until="2015-12-31", use.backtest=T)
## End(Not run)
}

