% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{neighbourhoods}
\alias{neighbourhoods}
\title{hexagonal bins}
\usage{
neighbourhoods(
  U,
  delta = (range(U$x, na.rm = TRUE)[2] - range(U$x, na.rm = TRUE)[1])/100
)
}
\arguments{
\item{U}{: a dataframe containing the numerical variables x and y}

\item{delta:}{controls the bin diameter}
}
\value{
a hexbin object  hexagonal bins
}
\description{
hexagonal bins
}
\examples{
# plot the hex bins of cumbria
data(U)
plot(neighbourhoods(U,.1))
plot(neighbourhoods(U,.01))
plot(neighbourhoods(U,.001))
}
