% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{newdist}
\alias{newdist}
\title{compute distances between new infected and exposed}
\usage{
newdist(
  closedistances = NULL,
  U,
  sicks,
  new.sicks = NULL,
  delta = 0.005,
  dist_areas = dist_areas_f(U, delta)
)
}
\description{
@param closedistances NULL, or a named list with 2 named elements: closedistances$ra, closedistances$id 
 @param U a data.frame with the variables hexagon (can be any bin identifier), x, y : coordinates, 
 @param sicks
 @param new.sicks
 @param delta a positive number : a threshold
 @param dist_areas: a function between 
 @return NULL, or a named list with 2 named elements: closedistances$ra, closedistances$id
}
\examples{
delta<-.005
sicks<-(1:nrow(U))[y=="sick"]
closedistances=newdist(NULL,U,sicks)
}
