% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_utils-Functions.R
\name{newStrategyFunction}
\alias{newStrategyFunction}
\title{Create Own Strategy}
\usage{
newStrategyFunction(name=NULL, file.path=getwd(), overwrite=F)
}
\description{
Creates a strategy functin template file. This file can be used as template for the development of customized strategies.
}
\examples{
##Not run:

# Creates a file myNewStrat.R at the specific file path
newStrategyFunction(name="myNewStrat", file.path=getwd(), overwrite=T)

##End(Not run)

}

