% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{performance}
\alias{performance}
\title{Get Strategy Performance}
\usage{
performance(object, which=NULL, of="portfolio", from=NULL, until=NULL, type="performance"
     , use.backtest=F
     , ...)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{of}{Performance to be extracted from assets separately or the portfolio performance.}

\item{from}{The date in character format \code{"yyyy-MM-dd"} or as date-object from which performance shall be returned If \code{NULL}, no restriction is made.}

\item{until}{The date in character format \code{"yyyy-MM-dd"} or as date-object until which performance shall be returned. If \code{NULL}, no restriction is made.}

\item{which}{Names or number of assets that should be included in performance. If a portfolio performance from only a subset of the assets is calculated, the weights are scaled accordingly.}

\item{type}{Which type of performance shall be returned. \code{performance} is the cumulative performance starting at \code{1}, \code{logReturns} to get logarithmic returns or \code{returns} for arithmetic returns.}

\item{use.backtest}{If \code{TRUE}, the signals from the backtesting output are considered for performance calculation. If \code{FALSE}, the signals from the initial strategy execution are used.}
}
\description{
Gets performance of an object of class \code{Strategy}.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get performance of MA(200)-Strategy
performance(myStrat.MA, from="2015-01-01", until="2015-12-31")

## Not run:
# Get backtest performance of MA(200)-Strategy
performance(myStrat.MA, from="2015-01-01", until="2015-12-31", use.backtest=T, type="logReturns")
## End(Not run)
}

