% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{plot}
\alias{plot}
\title{Plot of a \code{Strategy}-object}
\usage{
\S4method{plot}{Strategy}(x, from = NULL, until = NULL, which = NULL,
  main = NULL)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Calls the \code{plot} function that is defined within the strategy definition. E.g. MA-strategy's \code{plot} function is named \code{plot.MA()}.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Plot MA(200)-Strategy
plot(myStrat.MA, from="2015-01-01", until="2015-12-31", which=1)
}

