% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{polysmalldistmat}
\alias{polysmalldistmat}
\title{Compute distance matrix for a list of polygons}
\usage{
polysmalldistmat(
  list.poly,
  delta,
  gradients = cbind(c(0, 1), c(1, 0), c(1, 1), c(1, -1))
)
}
\arguments{
\item{list.poly}{a list of nx2 numeric matrices}
}
\value{
a  (n*(n-1)/2)x 3 matrix
}
\description{
Compute distance matrix for a list of polygons
}
\examples{
zz<-function(delta){
list.poly=plyr::alply(cbind(rep(0:8,9),rep(0:8,each=9))[sample(81,20),],1,function(x){
cbind(x[1]+c(0,0,.5,.5,0),x[2]+c(0,.5,.5,0,0))})
gradients=cbind(c(0,1),c(1,0),c(1,1),c(1,-1))
par(mfrow=c(1,1))
plot(do.call(rbind,list.poly),xlab="",yaxt="n",ylab="",cex=.1,main=paste0("Match polygons distant less than ",delta))
for(.poly in list.poly){points(.poly,type="l")}
X=polysmalldistmat(list.poly,delta)
for(i in 1:nrow(X)){
points(closestpointsontwopolygons(list.poly[[X[i,1]]],list.poly[[X[i,2]]]),col="red",type="l",lty=3)
}}
set.seed(1);zz(.5)
set.seed(1);zz(1)
set.seed(1);zz(2)
}
