% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{projpointonseg}
\alias{projpointonseg}
\title{computes the position of a projected point in the basis formed by a segment}
\usage{
projpointonseg(p, s)
}
\arguments{
\item{p}{a numeric vector of length 2}

\item{s}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}
}
\description{
computes the position of a projected point in the basis formed by a segment
}
\examples{
zz<-function(p){
s=matrix(c(0,1,0,0),2,2)
plot(s,type="l",xlim=c(-.5,1.5))
points(x=p[1],y=p[2] ,col="red",cex=.5)
points(projpointonseg(p,s)[1],projpointonseg(p,s)[2],col="red",cex=.5)
segments(x0 = p[1],y0=p[2],x1=projpointonseg(p,s)[1],y1=projpointonseg(p,s)[2],col="red")}
par(mfrow=c(3,3))
set.seed(1);replicate(9,zz(c(runif(1,-.5,1.5),runif(1,-1,1))))
}
