% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{projpointonseg_a}
\alias{projpointonseg_a}
\title{computes the position of a projected point in the basis formed by a segment}
\usage{
projpointonseg_a(p, s)
}
\arguments{
\item{p}{a numeric vector of length 2}

\item{s}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}
}
\description{
computes the position of a projected point in the basis formed by a segment
}
\examples{
data(Avo_fields,package="Strategy")
polygon1<-Avo_fields[1,]
A<-polygon1@polygons
B<-A[[1]]@Polygons[[1]]@coords
i=sample(nrow(B)-1,1)
s<-B[i:(i+1),]
p<-c(runif(1,min = 142.162,max=142.165),runif(1,min=-34.171,max=-34.167))
plot(B,type='l')
points(s,type="l",lwd=4,col="red")
points(x=p[1],y=p[2] ,col="red",cex=2)
points(projpointonseg(p,s)[1],projpointonseg(p,s)[2],col="red",cex=2)
segments(x0 = p[1],y0=p[2],x1=projpointonseg(p,s)[1],y1=projpointonseg(p,s)[2])
projpointonseg_a(p,s)
distpointtoseg(p,s)
dist(rbind(p,projpointonseg(p,s)))

zz<-function(p){
s=matrix(c(0,1,0,0),2,2)
plot(s,type="l",xlim=c(-.5,1.5))
points(x=p[1],y=p[2] ,col="red",cex=2)
points(projpointonseg(p,s)[1],projpointonseg(p,s)[2],col="red",cex=2)
segments(x0 = p[1],y0=p[2],x1=projpointonseg(p,s)[1],y1=projpointonseg(p,s)[2],col="red")
text(projpointonseg_a(p,s),-.8,paste0("a=",projpointonseg_a(p,s)))}
par(mfrow=c(2,2))
zz(c(-.5,1))
zz(0:1)
zz(c(.5,1))
zz(c(1.5,1))
}
