% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{ranges.gap}
\alias{ranges.gap}
\title{tells if two ranges overlap}
\usage{
ranges.gap(r1, r2)
}
\description{
tells if two ranges overlap
}
\examples{
zz<-function(){
x=sample(6,4,rep=F)
r1=x[1:2];r2=x[3:4]
plot(x,rep(0,4),main=ranges.gap(r1,r2),yaxt='n',xlab='',ylab='')
segments(x[1],0,x[2],0,col="blue")
segments(x[3],0,x[4],0,col="green")}
par(mfrow=c(2,2),oma=c(0,0,0,0))
set.seed(8);replicate(4,zz())
}
