% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{rangesbygradients_f}
\alias{rangesbygradients_f}
\title{Compute range of a polygon along certain gradients}
\usage{
rangesbygradients_f(
  .poly,
  gradients = -apply(cbind(c(0, 1), c(1, 0), c(1, 1), c(1, -1)), 2, function(x) {    
    x/(sqrt(sum(x^2))) })
)
}
\arguments{
\item{gradients}{a 2x n matrix each column representing a vector.}

\item{list.poly}{a list of nx2 numeric matrices}

\item{delta}{a positive number}
}
\description{
Compute range of a polygon along certain gradients
}
\examples{
a=c(1-1/sqrt(2),1/sqrt(2));.poly=cbind(1+c(0,0,a,1,1,a[2:1]),1+c(a,1,1,a[2:1],0,0))[c(1:8,1),]
plot(.poly,type='l',xlab='',ylab='')
rangesbygradients_f(.poly,gradients=cbind(c(0,1),c(1,0),c(1,-1)))
}
