% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{risktobeinfectedbydistancetoallinfectedunit}
\alias{risktobeinfectedbydistancetoallinfectedunit}
\title{Risk to be infected by many neighbours a neighbour at a certain distance}
\usage{
risktobeinfectedbydistancetoallinfectedunit(
  .dist,
  nI,
  .distriskhalf = 5 * 10^(-4),
  jumprisk = 10^-6
)
}
\arguments{
\item{.dist}{: a vector (distances)}

\item{.disthalfrisk}{: distance for which the risk is one half}

\item{nI:}{total number of infected}

\item{jumprisk:}{probability to be infected by one person, no matter how far he(she) is}
}
\value{
1-(prod(1-risktobeinfectedbydistancetooneinfectedunit(.dist,.distriskhalf))*(1-jumprisk)^nI)
}
\description{
Risk to be infected by many neighbours a neighbour at a certain distance
}
\examples{
#Risk to be ingfected 2 m from the victim when the 50\%risk distance is 1 m:
risktobeinfectedbydistancetooneinfectedunit(2,1)
}
