% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{risktobeinfectedbydistancetooneinfectedunit}
\alias{risktobeinfectedbydistancetooneinfectedunit}
\title{Risk to be infected by a neighbour at a distance x}
\usage{
risktobeinfectedbydistancetooneinfectedunit(.dist, .distriskhalf = 5 * 10^(-4))
}
\arguments{
\item{.dist}{: a distance}

\item{.disthalfrisk}{: distance for which the risk is one half}
}
\value{
exp(-.dist/(log(2)*.distriskhalf))
}
\description{
Risk to be infected by a neighbour at a distance x
}
\examples{
#Risk to be ingfected 2 m from the victim when the 50\%risk distance is 1 m:
risktobeinfectedbydistancetooneinfectedunit(2,1)
}
