% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{segment.intersect}
\alias{segment.intersect}
\title{computes the orientation of a triangle}
\usage{
segment.intersect(s1, s2)
}
\arguments{
\item{s1}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of an extreme point of the segment.}

\item{s2}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of an extreme point of the segment.}
}
\description{
computes the orientation of a triangle
}
\examples{
zz<-function(){
s1=matrix(sample(0:2,4,rep=T),2,2)
s2=matrix(sample(0:2,4,rep=T),2,2)
si<-segment.intersect(s1,s2)
s<-rbind(s1,s2)
plot(s,cex=.5,main=if(si){"Intersect"}else{"Disjoint"})
points(s1,type="l")
points(s2,type="l")
text(s[,1],s[,2],toupper(letters[1:4]),cex=2,col="red")

}
par(mfrow=c(3,3))
set.seed(1);replicate(9,zz())
}
