% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\docType{methods}
\name{sharpe}
\alias{sharpe}
\title{Get Sharpe Ratio of Performance}
\usage{
sharpe(object, rf=0, of="portfolio", which=NULL, from=NULL, until=NULL, scaling.periods=1
     , use.backtest=F
     , ...)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{rf}{Risk free rate in decimal, e.g. \code{rf=0.01} equals \code{1 percent}.}

\item{of}{Sharpe ratio to be calculated for assets separately or the portfolio sharpe.}

\item{from}{The date in character format \code{"yyyy-MM-dd"} or as date-object from which performance shall be considered. If \code{NULL}, no restriction is made.}

\item{until}{The date in character format \code{"yyyy-MM-dd"} or as date-object until which performance shall be considered. If \code{NULL}, no restriction is made.}

\item{which}{Names or number of assets that should be included in calculation.}

\item{scaling.periods}{Vector with annualization factors for sharpe ratio calculation. E.g. if data is monthly, scaling.periods is to be set to 12 for the respective .}

\item{use.backtest}{If \code{TRUE}, the performance of the backtesting output is considered for sharpe ratio calculation. If \code{FALSE}, the performance of the initial strategy execution are used.}
}
\description{
Get the sharpe ratio of the performance of an object of class \code{Strategy}.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)

# Get sharpe of MA(200)-Strategy portfolio
sharpe(myStrat.MA, from="2015-01-01", until="2015-12-31")

## Not run:
# Get backtest annualized sharpe of MA(200)-Strategy (daily data = 252 trading days)
sharpe(myStrat.MA, from="2015-01-01", until="2015-12-31", use.backtest=T, scaling.periods=252)
## End(Not run)
}

