% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyrisk2.R, R/shinyrisk2_sav.R
\name{Shiny.FieldLevelType2risk}
\alias{Shiny.FieldLevelType2risk}
\title{Shiny app - Simulation of infection and detection process}
\usage{
Shiny.FieldLevelType2risk(Fields0 = get(data(Fake, package = "Strategy")))

Shiny.FieldLevelType2risk(Fields0 = get(data(Fake, package = "Strategy")))
}
\arguments{
\item{Fields}{a Spatial Data frame with polygons}

\item{N}{population size}

\item{m}{number of infected}

\item{n0}{number of bulks in sample}

\item{n1}{number of trees in a bulk}

\item{n2}{number of colected leaves per sampled tree}

\item{beta}{risk at the leaf level, or 1-proportion of leaves with detectable RNA in an infected tree}
}
\value{
an integer, number of required leaves

an integer, number of required leaves
}
\description{
Shiny app - Simulation of infection and detection process

Shiny app - Simulation of infection and detection process
}
\examples{
requirednumberoftreeswithR(1237,6,1,1,0.63,0.05)
requirednumberoftreeswithoutR(1237,6,1,1,0.63,0.05)
riskwithoutR(1237,6,1236,1,0.63)
requirednumberoftreeswithoutR(1237,6,1,10,0.63,0.05)
requirednumberoftreeswithoutR(1237,6,1,10,0,0.05)
requirednumberoftreeswithR(1237,6,1,1,0.63,0.05)
requirednumberoftreeswithoutR(1237,6,1,1,0.63,0.05)
riskwithoutR(1237,6,1236,1,0.63)
requirednumberoftreeswithoutR(1237,6,1,10,0.63,0.05)
requirednumberoftreeswithoutR(1237,6,1,10,0,0.05)
}
