% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_class-Strategy.R
\name{Strategy}
\alias{Strategy}
\title{Create Strategy Object}
\usage{
Strategy(assets, stratFUN.name
    assetValueType = "price", weights = NULL, indicators = NULL,
    stratFUN.parameters = NULL, costs = NULL,
    printSteps = F)
}
\arguments{
\item{assets}{Time series of class \code{xts} of asset values in either price or log return form on
which the strategy function shall be applied. This is the portfolio of assets.}

\item{stratFUN.name}{The name of the strategy that should be applied. This can be either a
predefined strategy like MA or EWMA or a self-written function in
which case the full path to the function file to be called must be supplied.}

\item{assetValueType}{Assets can be passed as prices or log returns. In order to identify the
asset value types, either one of the types has to be selected.}

\item{weights}{The portfolio weights for the given assets as time series (dynamic) or numerical (constant) weights.}

\item{indicators}{A list of indicators that might be used within customized strategies. It is recommended to pass a named list.}

\item{stratFUN.parameters}{The list of parameters and their values required by the strategy function selected with parameter stratFUN.name.}

\item{costs}{The trading costs as constant costs per trade. This feature is not deployed yet and will be part of the bachelor thesis.}

\item{printSteps}{This is a feature used mainly for debugging the constructor function in order to localize where unspecified errors occur. If set to true, the different steps run within the constructor is printed to the console.}
}
\description{
Creates an object of class \code{Strategy} with the given portfolio data and strategy-function.
}
\examples{
# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, stratFUN.name="MA", stratFUN.parameters=params)
}

