% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures1..3.R
\name{subsets}
\alias{subsets}
\title{All ordered n-sized subsets of 1...N 
n}
\usage{
subsets(
  n,
  N,
  subsets_1 = if (n >= 2) {     subsets(n - 1, N) } else {     NULL }
)
}
\description{
All ordered n-sized subsets of 1...N 
n
}
\examples{
subsets(1,10)
subsets(2,10)
subsets(3,10)
subsets(8,10)
subsets(9,10)
subsets(10,10)
}
