% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{triangleorientation}
\alias{triangleorientation}
\title{computes the orientation of a triangle}
\usage{
triangleorientation(s)
}
\arguments{
\item{s}{a 3x2 numeric matrix, representing a triangle. Each row of the matrix are the coordinates of an extreme point of the triangle.}
}
\description{
computes the orientation of a triangle
}
\examples{
zz<-function(p){
s=matrix(sample(0:2,6,rep=T),3,2)
plot(s[c(1:3,1),],type="l",main=(if(triangleorientation(s)==1){"+"}else{if(triangleorientation(s)==-1){"-"}else{"aligned"}}))
text(s[,1],s[,2],toupper(letters[1:3]),cex=2,col="red")

}
par(mfrow=c(3,3))
replicate(9,zz(c(sample(-2:3,1),sample(-2:2,1))))
}
