% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{dist_areas_f}
\alias{dist_areas_f}
\title{Distances between hexagonal bins}
\usage{
dist_areas_f(
  U,
  delta = (range(U$x)[2] - range(U$x)[1])/100,
  h = neighbourhoods(U, delta)
)
}
\arguments{
\item{U}{: a dataframe containing the numerical variables x and y and preferable hexagon}

\item{delta}{: needed if hexagon is not a variable of U: bins will be recomputed}
}
\value{
a named matrix
}
\description{
Distances between hexagonal bins
}
\examples{
data(U) 
dist_areas_f(U)[1:3,1:3]
dist_areas_f(U,0.03)[1:3,1:3]
}
