% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{distpolytopoly}
\alias{distpolytopoly}
\title{Computes distance between two polygons}
\usage{
distpolytopoly(poly1, poly2)
}
\arguments{
\item{poly1}{a polygon (a n x 2 numerical matrix)}

\item{poly2}{a polygon (a n x 2 numerical matrix)}
}
\value{
a positive number, the distance between the two polygons
}
\description{
Computes distance between two polygons
}
\examples{
polys=lapply(c(0:1),function(x){
cbind(c(x,x,x+.5,x+.5,x),c(0,1,1,0,0))})
par(mfrow=c(1,1))
plot(do.call(rbind,polys),xlab="",yaxt="n")
for(.poly in polys){segments(x0 = .poly[-5,1],y0 = .poly[-5,2],.poly[-1,1],.poly[-1,2])}
distpolytopoly(polys[[1]],polys[[2]])
polys=lapply(c(1:2),function(x){
cbind(c(-x,-x,x,x,-x),c(-x,x,x,-x,-x))})
par(mfrow=c(1,1))
plot(do.call(rbind,polys),xlab="",yaxt="n")
for(.poly in polys){segments(x0 = .poly[-5,1],y0 = .poly[-5,2],.poly[-1,1],.poly[-1,2])}
distpolytopoly(polys[[1]],polys[[2]])
polys=lapply(c(1:2),function(x){
cbind(c(-2,-2,2,2,-2),c(-1,1,1,-1,-1))[,c(x,(1:2)[-x])]})
par(mfrow=c(1,1))
plot(do.call(rbind,polys),xlab="",yaxt="n")
for(.poly in polys){segments(x0 = .poly[-5,1],y0 = .poly[-5,2],.poly[-1,1],.poly[-1,2])}
distpolytopoly(polys[[1]],polys[[2]])
}
