% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{distsegmenttopoly}
\alias{distsegmenttopoly}
\title{Distance between a segment and a polygon}
\usage{
distsegmenttopoly(s, .poly)
}
\arguments{
\item{s}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}

\item{.poly}{a polygon (a nx2 matrix, each line is a point)}
}
\description{
Distance between a segment and a polygon
}
\examples{
zz<-function(){
.poly<-matrix(sample(0:6,6,T),3,2)[c(1:3,1),]
s<-matrix(sample(0:6,6,T),2,2)
plot(rbind(.poly,s),xlab="",yaxt="n")
points(.poly,type="l");points(s,type="l")
points(closestpointsontwopolygons(s,.poly),lty=3,col="red")}

}
