% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getBacktestSetup}
\alias{getBacktestSetup}
\alias{getBacktestSetup,Strategy-method}
\title{Get backtest parameter values from \code{Strategy}-object}
\usage{
getBacktestSetup(object)

\S4method{getBacktestSetup}{Strategy}(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Gets the backtest parameter values of an object of class \code{Strategy} that were used for backtesting the strategy. This includes the information about the parameters,
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Get backtest setup from MA(200)-Strategy
getBacktestSetup(myStrat.MA)

##End(Not run)
}
