% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getIndicators}
\alias{getIndicators}
\title{Get indicators from \code{Strategy}-object}
\usage{
getIndicators(object, from=NULL, until=NULL, which=NULL)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{which}{Names or list-number of indicators that should be included. If \code{NULL}, all indicators are returned.}
}
\description{
Gets the indicators data of an object of class \code{Strategy} that was used within strategy calculation.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
indicators <- list(volume=volume) # example: volume of trades
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params, indicators=indicators)

# Get indicator data from MA(200)-Strategy
getIndicators(myStrat.MA, from="2015-01-01", until="2015-12-31")

##End(Not run)
}

