% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getParameters}
\alias{getParameters}
\alias{getParameters,Strategy-method}
\title{Get strategy function parameters from \code{Strategy}-object}
\usage{
getParameters(object, use.backtest = FALSE)

\S4method{getParameters}{Strategy}(object, use.backtest = FALSE)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{use.backtest}{If set to \code{TRUE}, the calibrated parameters of the backtest are returned. Requires \code{\link{backtest}} to be  executed first.}
}
\description{
Gets the strategy function parameters of an object of class \code{Strategy} that were used for strategy calculation.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Get parameters from MA(200)-Strategy
getParameters(myStrat.MA)

##End(Not run)
}
