% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getSignals}
\alias{getSignals}
\title{Get trading signals from \code{Strategy}-object}
\usage{
getSignals(object, use.backtest=F, which=NULL)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{use.backtest}{If set to \code{TRUE}, the signals of the backtest are returned. Requires \code{\link{backtest}} to be  executed first.}

\item{which}{Names or column-number of assets that should be returned. If \code{NULL}, all prices are returned.}
}
\description{
Gets the trading signals of an object of class \code{Strategy} that were output from strategy calculation.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Get signals from MA(200)-Strategy
getSignals(myStrat.MA) # all signals returned
getSignals(myStrat.MA which=c(1,2), use.backtest=T) # backtest signals for first two assets returned

##End(Not run)
}

