% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-GET.R
\docType{methods}
\name{getStratFUN}
\alias{getStratFUN}
\alias{getStratFUN,Strategy-method}
\title{Get strategy function from \code{Strategy}-object}
\usage{
getStratFUN(object)

\S4method{getStratFUN}{Strategy}(object)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}
}
\description{
Gets the strategy function of an object of class \code{Strategy} that was used for strategy calculation.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Get strategy function from MA(200)-Strategy
MA.FUN <- getStratFUN(myStrat.MA)

##End(Not run)
}
