% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{newdist}
\alias{newdist}
\title{compute distances between new infected and exposed}
\usage{
newdist(
  closedistances = NULL,
  U,
  sicks,
  new.sicks = NULL,
  delta = 0.005,
  dist_areas = dist_areas_f(U, delta)
)
}
\arguments{
\item{closedistances}{NULL, or a named list with 2 named elements: closedistances$ra, closedistances$id}

\item{U}{a data.frame with the variables hexagon (can be any bin identifier), x, y : coordinates,}

\item{sicks}{a vector of integers}

\item{new.sicks}{a vector of integers}

\item{delta}{a positive number : a threshold}

\item{dist_areas:}{a function between}
}
\value{
NULL, or a named list with 2 named elements: closedistances$ra, closedistances$id
}
\description{
compute distances between new infected and exposed
}
\examples{
data(UE,package="Strategy")
delta<-.005
sicks<-(1:nrow(UE))[UE$I001=="sick"]
closedistances=newdist(NULL,UE,sicks)
do.call(cbind,closedistances)[1:3,]
}
