% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-PLOT.R
\docType{methods}
\name{plot}
\alias{plot}
\title{Plot of a \code{Strategy}-object}
\usage{
\S4method{plot}{Strategy}(x, from = NULL, until = NULL,
  which.assets = NULL, which.filters = NULL, which.indicators = NULL,
  main = NULL, show.signals = T, include.costs = T, ...)
}
\arguments{
\item{x}{An object of class \code{Strategy}.}

\item{from}{From date that chart is to be plotted.}

\item{until}{Until date that chart is to be plotted.}

\item{which.assets}{Which assets shall be plotted (each one will result in single plot)}

\item{which.filters}{Which filters shall be added to price plot. Default value \code{NULL} will return all filters from the strategy.}

\item{which.indicators}{Which indicators shall be added to indicator plot. Default value \code{NULL} will return all filters from the strategy. If \code{"none"}, no indicator is plotted and indicator area is not shown.}

\item{main}{The main title of the plot.}

\item{include.costs}{If \code{FALSE}, the fixed and relative trading costs are NOT considered for performance calculation. Default value is \code{TRUE}. As default values for costs are \code{0}, this argument is redundant if no costs are given.}
}
\description{
Calls the \code{plot} function that is defined within the strategy definition. E.g. MA-strategy's \code{plot} function is named \code{plot.MA()}.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Plot first asset of MA(200)-Strategy
plot(myStrat.MA, from="2015-01-01", until="2015-12-31", which=1)

##End(Not run)
}

