% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_methods-PLOT.R
\docType{methods}
\name{plotWeights}
\alias{plotWeights}
\alias{plotWeights,Strategy-method}
\title{Plot Strategy Weights}
\usage{
plotWeights(object, from = NULL, until = NULL, ...)

\S4method{plotWeights}{Strategy}(object, from = NULL, until = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{Strategy}.}

\item{from}{The date in character format \code{"yyyy-MM-dd"} or as date-object from which weights shall be plotted. If \code{NULL}, the start date of the weights is used.}

\item{until}{The date in character format \code{"yyyy-MM-dd"} or as date-object until which weights shall be plotted. If \code{NULL}, the end date of the performances is used.}

\item{...}{Currently not active.}
}
\description{
Plots the weights of the portfolio of an object of class \code{Strategy}.
}
\examples{
##Not run:

# MA(200)-Strategy
params <- list(k=200)
myStrat.MA <- Strategy(assets=assets, strat="MA", strat.params=params)

# Plot MA(200)-Strategy weights
plotWeights(myStrat.MA)

##End(Not run)
}
