% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{polydistmat}
\alias{polydistmat}
\title{Compute distance matrix for a list of polygons}
\usage{
polydistmat(list.poly)
}
\arguments{
\item{list.poly}{a list of nx2 numeric matrices}
}
\value{
a  (n*(n-1)/2)x 3 matrix
}
\description{
Compute distance matrix for a list of polygons
}
\examples{
polys=lapply(c(0:3,5:7),function(x){
cbind(c(x,x,x+.5,x+.5,x),c(0,1,1,0,0))})
par(mfrow=c(1,1))
plot(do.call(rbind,polys),xlab="",yaxt="n")
for(.poly in polys){segments(x0 = .poly[-5,1],y0 = .poly[-5,2],.poly[-1,1],.poly[-1,2])}
polydistmat(polys)
#data(Avo_fields,package="Strategy")
#polygons<-Avo_fields
#MM<-polydistmat(extractpolygonsaslist(Avo_fields))
#parallel::detectCores()
#save(MM,file=file.path(Mydirectories::googledrive.directory(),"Travail/Recherche/Travaux/Epidemiologie/Strategy/data/MM.rda"))
}
