% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{projpointonseg_a}
\alias{projpointonseg_a}
\title{computes the position of a projected point in the basis formed by a segment}
\usage{
projpointonseg_a(p, s, method = "euclidean")
}
\arguments{
\item{p}{a numeric vector of length 2}

\item{s}{a 2x2 numeric matrix, representing a segment. Each row of the matrix are the coordinates of a extreme point of the segment.}
}
\description{
computes the position of a projected point in the basis formed by a segment
}
\examples{
zz<-function(p){
s=matrix(c(0,1,0,0),2,2)
plot(s,type="l",xlim=c(-.5,1.5),
xlab="",ylab="",
xaxt='n',yaxt='n')
points(x=p[1],y=p[2] ,col="red",cex=2)
points(closestpointonseg(p,s)[1],closestpointonseg(p,s)[2],col="red",cex=2)
segments(x0 = p[1],y0=p[2],x1=closestpointonseg(p,s)[1],y1=closestpointonseg(p,s)[2],col="red")
text(projpointonseg_a(p,s),-.8,paste0("a=",projpointonseg_a(p,s)))}
par(oma=c(0,0,0,0),mfrow=c(1,4))
zz(c(-.5,1))
zz(0:1)
zz(c(.5,1))
zz(c(1.5,1))
}
