% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{ranges.gap}
\alias{ranges.gap}
\title{Interval length between two ranges (0 if they overlap)}
\usage{
ranges.gap(r1, r2)
}
\arguments{
\item{r1}{a range a length 2 numerical vector}

\item{r2}{a range a length 2 numerical vector}
}
\value{
TRUE if two ranges overlap
}
\description{
Interval length between two ranges (0 if they overlap)
}
\examples{
par(mfrow=c(1,4),oma=c(0,0,0,0))
set.seed(10);replicate(4,(function(){
r1<-sample(1:8,2);r2<-sample(1:5,2)
plot(cbind(c(r1,r2),0),yaxt='n',xlab='',ylab='',
main=paste0("Gap: ",ranges.gap(r1,r2)))
points(cbind(r1,0),type="l",lwd=3)
points(cbind(r2,0),type="l",col="red")
})())
}
